Collapse schema instances into models
